+++ pages/soon-in-theater.tsx
@@ -1,0 +1,8 @@
+import MoviesList from '../components/movies-list';
+
+export default function InTheater() {
+
+  return (
+    <MoviesList pageTitle={'Em breve nos cinemas'} urlPrefix={'upcoming'} />
+  )
+}
